Move echem block to its own app submodule
