+++ pydatalab/pydatalab/apps/echem/__init__.py
@@ -1,0 +1,3 @@
+from .blocks import CycleBlock
+
+__all__ = ("CycleBlock",)
